# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subscriptions/me": {
      "get": {
        "description": "Active plan, status and period.",
        "operationId": "SubscriptionsController_getCurrent_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get your subscription status",
        "tags": ["Subscriptions"]
      },
      "patch": {
        "description": "Switches to another active plan immediately. The previous subscription is kept as CANCELED history. Usage already counted in the current period carries over; the new limit applies at once.",
        "operationId": "SubscriptionsController_changePlan_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the request is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_FAILED",
                  "message": "Request validation failed",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or not owned by the caller)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 404,
                  "code": "PLAN_NOT_FOUND",
                  "message": "Plan \"gold\" does not exist",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state of a resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 409,
                  "code": "PLAN_ALREADY_ACTIVE",
                  "message": "You are already on the Premium plan",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upgrade or downgrade your plan",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/me/history": {
      "get": {
        "description": "Newest first.",
        "operationId": "SubscriptionsController_history_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List your subscription history",
        "tags": ["Subscriptions"]
      }
    },
    "/api/v1/subscriptions/me/usage": {
      "get": {
        "description": "Usage of the plan allowance in the current UTC day or month, and when it resets.",
        "operationId": "SubscriptionsController_getUsage_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get remaining requests",
        "tags": ["Subscriptions"]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "REST API for the EchoGPT multi-AI chat Chrome extension.\n\n**Authentication:** call `POST /auth/login`, then send `Authorization: Bearer <accessToken>`.\nRefresh with `POST /auth/refresh` before the access token expires.\n\n**Errors** always use the `ErrorResponseDto` envelope with a stable `code`.\n**Tracing:** send `x-request-id` to correlate logs; it is echoed back on every response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable error code"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Extra context, e.g. field-level validation errors",
            "example": null
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource/0192f0c4-8a3e-7c1e-9d2b-5b1f3a9c2e10"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
          }
        },
        "required": ["statusCode", "code", "message", "path", "timestamp", "requestId"]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "premium"
          },
          "name": {
            "type": "string",
            "example": "Premium"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Higher daily allowance for power users"
          },
          "priceCents": {
            "type": "number",
            "description": "Price in the smallest currency unit",
            "example": 999
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "requestLimit": {
            "type": "number",
            "description": "Requests allowed per period",
            "example": 500
          },
          "limitPeriod": {
            "enum": ["DAILY", "MONTHLY"],
            "type": "string",
            "example": "DAILY"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "priceCents",
          "currency",
          "requestLimit",
          "limitPeriod"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": ["ACTIVE", "CANCELED", "EXPIRED"],
            "type": "string",
            "example": "ACTIVE"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanResponseDto"
          },
          "currentPeriodStart": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "canceledAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "plan",
          "currentPeriodStart",
          "currentPeriodEnd",
          "canceledAt",
          "createdAt"
        ]
      },
      "UsageResponseDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "example": "free"
          },
          "period": {
            "enum": ["DAILY", "MONTHLY"],
            "type": "string",
            "example": "DAILY"
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "used": {
            "type": "number",
            "example": 7
          },
          "remaining": {
            "type": "number",
            "example": 13
          },
          "periodStart": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-26T00:00:00.000Z"
          },
          "resetsAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-27T00:00:00.000Z"
          }
        },
        "required": ["planCode", "period", "limit", "used", "remaining", "periodStart", "resetsAt"]
      },
...
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "pattern": "^[a-z0-9_-]+$",
            "example": "premium",
            "description": "Code of the target plan (see GET /plans)"
          }
        },
        "required": ["planCode"]
      },
...
    }
  }
}